#96 Refactor to template
