+++ .github/linters/.jscpd.json
