+++ .github/workflows/pytest-darwin.yaml
@@ -15,7 +15,7 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        os: [macos-13]
+        os: [macos-latest]
         # 3.8 and 3.9 are unsupported on the darwin runner
         python-version: ["3.10", "3.11"]
