+++ .gitignore
@@ -155,3 +155,5 @@ cython_debug/
 # VSCode
 .vscode/
 .history
+
+.settings
