+++ CONTRIBUTING.md
@@ -16,7 +16,7 @@ There are a number of ways you can contribute:
 
 If your contribution modifies the git repository, the following agreements must be established.
 
-*Note:*  License agreements are only needed for adding, modifying, and deleting artifacts kept within the repository.
+_Note:_ License agreements are only needed for adding, modifying, and deleting artifacts kept within the repository.
 In simple terms, license agreements are needed before pull requests can be accepted.
 A license agreement is not needed for submitting feature request, bug reporting, or other project management.
