+++ README.md
@@ -23,6 +23,8 @@ implmentations of the Senzing Python SDK are derived.
 [![PEP8 Badge]][PEP8]
 [![PyPI version Badge]][PyPi version]
 [![Downloads Badge]][Downloads]
+[![License Badge]][License]
+[![Coverage Badge]][Coverage]
 
 ## Overview
 
@@ -43,6 +45,8 @@ implmentations of the Senzing Python SDK are derived.
 [Downloads]: https://pepy.tech/project/senzing-abstract
 [Errors]: docs/errors.md
 [Examples]: docs/examples.md
+[License Badge]: https://img.shields.io/badge/License-Apache2-brightgreen.svg
+[License]: https://github.com/senzing-garage/template-go/blob/main/LICENSE
 [PEP8 Badge]: https://img.shields.io/badge/code%20style-pep8-orange.svg
 [PEP8]: https://www.python.org/dev/peps/pep-0008/
 [PyPI version Badge]: https://badge.fury.io/py/senzing-abstract.svg
