+++ pyproject.toml
@@ -4,3 +4,67 @@ requires = [
     "wheel"
 ]
 build-backend = "setuptools.build_meta"
+
+[dependency-groups]
+test = ["pytest", "coverage"]
+
+
+[project]
+name = "senzing-core"
+authors = [
+    {name = "Senzing", email = "support@senzing.com"},
+]
+readme = "README.md"
+requires-python = ">=3.8"
+classifiers = [
+    "Programming Language :: Python :: 3",
+    "License :: OSI Approved :: Apache Software License",
+    "Operating System :: OS Independent"
+]
+dynamic = ["version", "description"]
+
+dependencies = [
+    "orjson==3.10.9",
+]
+
+[tool.bandit]
+skips = ["B101"]
+
+[tool.flake8]
+extend-ignore = ["E203", "E501", "E704", "W503"]
+max-line-length = 88
+
+[tool.isort]
+profile = "black"
+src_paths = ["examples", "src", "tests"]
+
+[[tool.mypy.overrides]]
+module = "senzing_abstract.szengineflags.*"
+ignore_missing_imports = "true"
+warn_unused_ignores = "false"
+
+[[tool.mypy.overrides]]
+module = "pytest_schema.*"
+ignore_missing_imports = "true"
+
+[tool.pylint]
+ignored-argument-names = "args|kwargs"
+disable = [
+    "duplicate-code",
+    "line-too-long",
+    "missing-module-docstring",
+    "too-many-arguments",
+    "too-many-function-args",
+    "too-many-instance-attributes",
+    "too-many-positional-arguments",
+    "too-many-public-methods"
+]
+ignore = [
+    "__init__.py",
+    "docs/source/conf.py"
+]
+notes = [
+    "FIXME"
+]
+
+
